+++ changelog.md
@@ -7,4 +7,9 @@
 ## 1.0.1
 
 * Automatically find relevent images for a GroupedProduct
-* Automatically merge all images from a group 
+* Automatically merge all images from a group
+
+## 1.0.2
+
+* Filter products from a group that are disabled
+* Disable items that are out of stock (if stock keeping is installed)